requirements.txt
asgiref==3.9.0
bleach==6.2.0
diff-match-patch==20241021
Django==5.2.4
django-cors-headers==4.7.0
django-import-export==4.3.8
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-api-logger==1.1.19
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
webencodings==0.5.1
python-dotenv==1.0.0
django-environ==0.11.2
